requirements.txt
anthropic
deepeval
pytest
pytest-asyncio
python-dotenv